move pg prefix in file name
