+++ pgdrive/component/blocks/curve.py
@@ -4,7 +4,7 @@ from pgdrive.component.blocks.pg_block import PGBlock
 from pgdrive.component.blocks.create_block_utils import CreateAdverseRoad, CreateRoadFrom, create_bend_straight
 from pgdrive.constants import LineType
 from pgdrive.component.road.road import Road
-from pgdrive.utils.pg_space import ParameterSpace, Parameter, BlockParameterSpace
+from pgdrive.utils.space import ParameterSpace, Parameter, BlockParameterSpace
 
 
 class Curve(PGBlock):
