+++ pgdrive/component/blocks/straight.py
@@ -3,7 +3,7 @@ from pgdrive.component.blocks.create_block_utils import ExtendStraightLane, Crea
 from pgdrive.constants import LineType
 from pgdrive.component.lane.straight_lane import StraightLane
 from pgdrive.component.road.road import Road
-from pgdrive.utils.pg_space import ParameterSpace, Parameter, BlockParameterSpace
+from pgdrive.utils.space import ParameterSpace, Parameter, BlockParameterSpace
 
 
 class Straight(PGBlock):
